+++ ai/core/src/main/kotlin/org/http4k/ai/model/ApiKey.kt
@@ -1,8 +1,9 @@
 package org.http4k.ai.model
 
+import dev.forkhandles.values.Maskers.hidden
 import dev.forkhandles.values.NonBlankStringValueFactory
 import dev.forkhandles.values.StringValue
 
-class ApiKey private constructor(value: String) : StringValue(value) {
+class ApiKey private constructor(value: String) : StringValue(value, hidden()) {
     companion object : NonBlankStringValueFactory<ApiKey>(::ApiKey)
 }