+++ connect/amazon/core/src/main/kotlin/org/http4k/connect/amazon/core/model/SecretAccessKey.kt
@@ -1,8 +1,9 @@
 package org.http4k.connect.amazon.core.model
 
+import dev.forkhandles.values.Maskers.hidden
 import dev.forkhandles.values.NonBlankStringValueFactory
 import dev.forkhandles.values.StringValue
 
-class SecretAccessKey private constructor(value: String) : StringValue(value) {
+class SecretAccessKey private constructor(value: String) : StringValue(value, hidden()) {
     companion object : NonBlankStringValueFactory<SecretAccessKey>(::SecretAccessKey)
 }