+++ connect/amazon/core/src/main/kotlin/org/http4k/connect/amazon/core/model/WebIdentityToken.kt
@@ -1,10 +1,11 @@
 package org.http4k.connect.amazon.core.model
 
+import dev.forkhandles.values.Maskers.hidden
 import dev.forkhandles.values.NonEmptyStringValueFactory
 import dev.forkhandles.values.StringValue
 import java.io.File
 
-class WebIdentityToken private constructor(value: String) : StringValue(value) {
+class WebIdentityToken private constructor(value: String) : StringValue(value, hidden()) {
     companion object : NonEmptyStringValueFactory<WebIdentityToken>(::WebIdentityToken) {
         fun of(file: File) = of(file.readText())
     }