+++ connect/gitlab/client/src/main/kotlin/org/http4k/connect/gitlab/GitLabToken.kt
@@ -1,8 +1,9 @@
 package org.http4k.connect.gitlab
 
+import dev.forkhandles.values.Maskers.hidden
 import dev.forkhandles.values.NonBlankStringValueFactory
 import dev.forkhandles.values.StringValue
 
-class GitLabToken private constructor(value: String) : StringValue(value) {
+class GitLabToken private constructor(value: String) : StringValue(value, hidden()) {
     companion object : NonBlankStringValueFactory<GitLabToken>(::GitLabToken)
 }